requirements.txt
pycryptodome>=3.15.0
cryptography>=3.4.8
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
flask>=3.0.0
flask-socketio>=5.3.6
scikit-learn>=1.3.2
scipy>=1.11.4
pandas>=2.1.4 